Improve locale build

+++ packages/locale/build.sh
@@ -5,6 +5,9 @@ if [[ "$PWD" != *"/locale" ]]; then
     exit 1
 fi
 
-podman run --rm -v "$PWD/../..:/usr/src/based-db" based-db-clibs-build-linux_aarch64 /bin/bash -c "cd /usr/src/based-db/packages/locale && make -j4"
-podman run --rm -v "$PWD/../..:/usr/src/based-db" based-db-clibs-build-linux_x86_64 /bin/bash -c "cd /usr/src/based-db/packages/locale && make -j4"
-podman rm --all
+if test "$OSTYPE" == "linux-gnu" && test "$1" != "release"; then
+    make -j4
+else
+    podman run --rm -v "$PWD/../..:/usr/src/based-db" based-db-clibs-build-linux_aarch64 /bin/bash -c "cd /usr/src/based-db/packages/locale && make -j4"
+    podman run --rm -v "$PWD/../..:/usr/src/based-db" based-db-clibs-build-linux_x86_64 /bin/bash -c "cd /usr/src/based-db/packages/locale && make -j4"
+fi
